Refactor report.py
This is the first refactor. I wanted access to scores and error
rates in notebooks. This refactor uncovered a couple mistakes:
the score function used len(hits) when it should have used
sum(hits). And the 'all' category was included in scatter plots
when it should not be considered a category.